Moved AppKey class to a separated file
