+++ iot_api/user_api/models/__init__.py
@@ -17,3 +17,4 @@ from iot_api.user_api.models.Tag import Tag
 from iot_api.user_api.models.DeviceToTag import DeviceToTag
 from iot_api.user_api.models.GatewayToTag import GatewayToTag
 from iot_api.user_api.models.NotificationAssetTag import NotificationAssetTag
+from iot_api.user_api.models.AppKey import AppKey
